pyproject.toml
[tool.poetry]
name = "summariser"
version = "0.1.0"
description = ""
authors = ["Franklin Obasi <franklinobasy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
predibase = "^2024.11.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
